V4L2 Consolidation
This commit includes the following revisions:
1.  Revise function names for video2.c to be prefixed with v4l2
2.  Revise function names for video_common.c to be prefixed with vid
3.  Consolidate the v4l2 functions into video2.c
4.  Eliminate hard coded parameter to vloopback.
5.  Eliminate some of the GREY pallete conditions since GREY it
    is not possible to have that selection.
6.  Changes to video_freebsd to conform with new naming
